- 为各 MenuFlyout 提取基类，以减少重复代码。 - 修复隐藏专辑与歌曲的多选按钮时，MenuFlyoutSeparator 未隐藏的问题。
