+++ src/MonsterSiren.Uwp/Controls/MenuFlyouts/SongListMenuFlyout.xaml
@@ -1,4 +1,4 @@
-<local:WithAddToMenuFlyout
+<local:ListMenuFlyoutBase
     x:Class="MonsterSiren.Uwp.Controls.MenuFlyouts.SongListMenuFlyout"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -66,4 +66,4 @@
         controls:CommandParameterXamlHelper.SourceData="{x:Bind SourceData, Mode=OneWay}"
         Command="{x:Bind StopMultipleSelectionCommand}"
         Icon="Cancel" />
-</local:WithAddToMenuFlyout>
+</local:ListMenuFlyoutBase>
